moved temporary data to assets and disabled robots.txt when NUXT_APP_BASE_URL is set
